feat - add exceptions and logs

+++ src/main/java/org/example/ecommerce/service/OrderService.java
@@ -1,8 +1,12 @@
 package org.example.ecommerce.service;
 
 import jakarta.transaction.Transactional;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.example.ecommerce.dto.request.OrderRequestDTO;
 import org.example.ecommerce.dto.response.OrderResponseDTO;
+import org.example.ecommerce.infra.config.exception.OrderListIsEmptyException;
+import org.example.ecommerce.infra.config.exception.OrderNotFoundException;
 import org.example.ecommerce.model.Order;
 import org.example.ecommerce.model.Product;
 import org.example.ecommerce.reporitory.OrderRepository;
@@ -24,11 +28,24 @@ public class OrderService {
     @Autowired
     ProductRepository productRepository;
 
+    private static final Logger LOGGER = LogManager.getLogger();
+
     public OrderResponseDTO findById(UUID id) {
-        return orderRepository.findById(id).map(OrderResponseDTO::new).orElseThrow(() -> new RuntimeException("Order not found."));
+        LOGGER.info("Getting order...");
+        return orderRepository.findById(id).map(OrderResponseDTO::new).orElseThrow(() -> {
+            LOGGER.error("Order not found");
+            return new OrderNotFoundException("Order not found.");
+        });
     }
 
     public List<OrderResponseDTO> findAllOrders() {
+        LOGGER.info("Getting all orders...");
+
+        if(orderRepository.findAll().isEmpty()) {
+            LOGGER.error("Orders are empty.");
+            throw new OrderListIsEmptyException("Orders are empty!");
+        }
+
         return orderRepository.findAll().stream().map(OrderResponseDTO::new).toList();
     }
 
@@ -61,10 +78,19 @@ public class OrderService {
         order.setBuyer(orderRequestDTO.getBuyer());
         order.setExpectedDate(orderRequestDTO.getExpectedDate());
         order.setProducts(products);
+        LOGGER.info("Saving an order...");
         return orderRepository.save(order);
     }
 
     public void deleteOrder(UUID id) {
+        LOGGER.info("Deleting order...");
+
+        if(orderRepository.findById(id).isEmpty()) {
+            LOGGER.error("Order not found.");
+            throw new OrderNotFoundException("Order not found");
+        }
+
         orderRepository.deleteById(id);
+        LOGGER.info("Order deleted successfully.");
     }
 }
